// fe/src/lib/api/generated/model/resourceStatusValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type ResourceStatusValue = typeof ResourceStatusValue[keyof typeof ResourceStatusValue];


export const ResourceStatusValue = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED',
} as const;